requirements.txt
langchain_google_genai
crewai
crewai_tools
langchain_community
python-dotenv
markdown
ipython
matplotlib
seaborn
pandas